rename domain product catalog router
